+++ src/agent/balancer.js
@@ -160,8 +160,10 @@ var Balancer = {
     if (this.isRunnig()) {
       log.debug("call to stop balancer");
       return Tools.async_status("balancer", this, function* (change_status) {
-        yield this._stop_system('balancer-redirect', change_status);
-        yield this._stop_system('dns', change_status);
+        yield Q.all([
+          this._stop_system('balancer-redirect', change_status),
+          this._stop_system('dns', change_status),
+        ]);
         yield this._stop_sub_service("hipache", change_status);
         yield this._stop_sub_service("memcached", change_status);
       });